// src/gen/ride/wallet/v1alpha1/wallet_service.ts
export interface Account {
    /**
     * @generated from protobuf field: string account_id = 1;
     */
    accountId: string;
    /**
     * @generated from protobuf field: string uid = 2;
     */
    uid: string;
    /**
     * @generated from protobuf field: int32 balance = 3;
     */
    balance: number;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp create_time = 7;
     */
    createTime?: Timestamp;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp update_time = 8;
     */
    updateTime?: Timestamp;
}
// ...
class Account$Type extends MessageType<Account> {
    constructor() {
        super("ride.wallet.v1alpha1.Account", [
            { no: 1, name: "account_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "uid", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "balance", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 7, name: "create_time", kind: "message", T: () => Timestamp },
            { no: 8, name: "update_time", kind: "message", T: () => Timestamp }
        ]);
    }
    create(value?: PartialMessage<Account>): Account {
        const message = { accountId: "", uid: "", balance: 0 };
        globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this });
        if (value !== undefined)
            reflectionMergePartial<Account>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Account): Account {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string account_id */ 1:
                    message.accountId = reader.string();
                    break;
                case /* string uid */ 2:
                    message.uid = reader.string();
                    break;
                case /* int32 balance */ 3:
                    message.balance = reader.int32();
                    break;
                case /* google.protobuf.Timestamp create_time */ 7:
                    message.createTime = Timestamp.internalBinaryRead(reader, reader.uint32(), options, message.createTime);
                    break;
                case /* google.protobuf.Timestamp update_time */ 8:
                    message.updateTime = Timestamp.internalBinaryRead(reader, reader.uint32(), options, message.updateTime);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: Account, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string account_id = 1; */
        if (message.accountId !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.accountId);
        /* string uid = 2; */
        if (message.uid !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.uid);
        /* int32 balance = 3; */
        if (message.balance !== 0)
            writer.tag(3, WireType.Varint).int32(message.balance);
        /* google.protobuf.Timestamp create_time = 7; */
        if (message.createTime)
            Timestamp.internalBinaryWrite(message.createTime, writer.tag(7, WireType.LengthDelimited).fork(), options).join();
        /* google.protobuf.Timestamp update_time = 8; */
        if (message.updateTime)
            Timestamp.internalBinaryWrite(message.updateTime, writer.tag(8, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message ride.wallet.v1alpha1.Account
 */
export const Account = new Account$Type();